Initial Avionics Bay Designs
The initial designs of the av bay, to be refined once I can open the intrepid.iam file and calculate the exact dimensions. It will also need to be refined once we know what is going into it and how it needs to be mounted, along with the expected pressure and temperature of the bay.